Alterando estrutura para salvamento de senhas. Criptografia adicionada. Criado módulo para comparação de senhas usando Bcrypt. Alterando estrutura para comportar logins e autenticação. Instalado JWT no projeto para futuro uuso.
