+++ package.json
@@ -23,6 +23,7 @@
     "bcrypt": "5.1.0",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
+    "jsonwebtoken": "^9.0.0",
     "mysql2": "^3.2.0",
     "sequelize": "6.29.3",
     "swagger-ui-express": "^4.6.2"
